# Fixed slowness in unicodeTestIn script used by keyword template.