--- a/news/125.bugfix
+++ b/news/125.bugfix
@@ -1,0 +1,2 @@
+Fixed slowness in ``unicodeTestIn`` script used by keyword template.
+[maurits]