Fixes issue #125 for branch 1.9.x (Plone 4.3).

This contains performance tests, which are ignored by default, but can be activated by exporting a number:

```
$ export AT_UNICODETESTIN=3000
$ bin/test -s Products.Archetypes -m test_widgets -t test_unicodeTestIn
...
Testing unicodeTestIn with 3000 items, and 50 selected.
...
```

Timings before this fix:

```
First test part finished in 4.2931 seconds
Second test part finished in 12.6825 seconds
```

Timings after this fix:

```
First test part finished in 0.0373 seconds
Second test part finished in 0.2010 seconds
```

And with the fix and 30,000 items:

```
First test part finished in 0.0374 seconds
Second test part finished in 2.0301 seconds
```